pyproject.toml
[project]
name = "cli-llm-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "click>=8.2.1",
    "colorama>=0.4.6",
    "openai>=1.99.5",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "rich>=14.1.0",
    "tenacity>=9.1.2",
    "wikipedia>=1.4.0",
]
